Harden concurrency hot paths: memory-model safety, single-compute caches, overflow-safe round-robin
Five thread-safety/perf fixes surfaced by the audit, all behavior-preserving
(full Core suite green, 1606 tests):

- MessageHandlerDefinitionIndex: publish the built index and its version as a
  single immutable state object through one volatile reference. The prior
  lock-free fast path read a non-volatile reference + int, so a reader on a
  weak memory model (ARM64/Graviton) could see the published dictionary before
  its contents - or its matching version - became visible.
- MessageHandlerDefinitionLookUp.FindHandler: resolve the target version once
  instead of re-running version selection (and re-allocating the candidate
  array) inside the FirstOrDefault predicate for every handler - O(n^2) work
  and allocation per dispatch collapsed to O(n).
- CacheMessageHandlersFinder / CacheHttpEndpointFinder: replace the non-atomic
  null-coalescing-assignment first-call cache with a double-checked lock over a
  volatile field, so concurrent first calls no longer both run the
  (reflection-based) discovery.
- BoundedConcurrentDispatcher: reduce the round-robin counter through uint so
  it can't wrap to int.MinValue and produce a negative lane index after
  int.MaxValue enqueues.

+++ src/Benzene.SelfHost/BoundedConcurrentDispatcher.cs
@@ -95,9 +95,13 @@ public sealed class BoundedConcurrentDispatcher<T>
     /// <param name="cancellationToken">Cancels the enqueue.</param>
     public ValueTask EnqueueAsync(T item, CancellationToken cancellationToken)
     {
+        // Both paths reduce through uint: the round-robin counter increments without bound and would,
+        // after int.MaxValue enqueues on one dispatcher, wrap to int.MinValue - whose signed modulo is
+        // negative and would index _lanes[-x] (IndexOutOfRangeException). Casting to uint first keeps
+        // the index in [0, laneCount) across the wrap.
         var laneIndex = _keySelector != null
-            ? (int)((uint)_keySelector(item) % _lanes.Length)
-            : Interlocked.Increment(ref _roundRobinCounter) % _lanes.Length;
+            ? (int)((uint)_keySelector(item) % (uint)_lanes.Length)
+            : (int)((uint)Interlocked.Increment(ref _roundRobinCounter) % (uint)_lanes.Length);
 
         return _lanes[laneIndex].Writer.WriteAsync(item, cancellationToken);
     }
